requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
Flask-Cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
boto3==1.34.0
gunicorn==22.0.0
